[pkg/config] Move `pkg/config/seelog` to `pkg/config/internal/seelog`
