+++ pkg/config/log_test.go
@@ -10,7 +10,7 @@ import (
 	"bytes"
 	"testing"
 
-	seelogCfg "github.com/DataDog/datadog-agent/pkg/config/seelog"
+	seelogCfg "github.com/DataDog/datadog-agent/pkg/config/internal/seelog"
 	"github.com/cihub/seelog"
 	"github.com/stretchr/testify/assert"
 )
